Rename scudesc to api
