+++ azazel/api/views.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.models import User, Group
 from rest_framework import viewsets
-from azazel.scudesc.serializers import UserSerializer, GroupSerializer
+from azazel.api.serializers import UserSerializer, GroupSerializer
 
 
 class UserViewSet(viewsets.ModelViewSet):
