+++ azazel/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url, include
 from rest_framework import routers
-from azazel.scudesc import views
+from azazel.api import views
 
 router = routers.DefaultRouter()
 router.register(r'users', views.UserViewSet)
